refactor! Refactor and improve the API
