+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/agent/entity/graph/AIAgentNode.kt
@@ -1,4 +1,4 @@
-package ai.koog.agents.core.agent.entity
+package ai.koog.agents.core.agent.entity.graph
 
 import ai.koog.agents.core.agent.context.AIAgentContextBase
 import ai.koog.agents.core.agent.context.element.NodeInfoContextElement
