+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/agent/entity/graph/SubgraphMetadata.kt
@@ -1,4 +1,4 @@
-package ai.koog.agents.core.agent.entity
+package ai.koog.agents.core.agent.entity.graph
 
 /**
  * Represents metadata associated with a subgraph in an AI agent strategy graph.
