+++ agents/agents-core/src/commonMain/kotlin/ai/koog/agents/core/feature/handler/ExecuteNodeEventHandlerContext.kt
@@ -1,7 +1,7 @@
 package ai.koog.agents.core.feature.handler
 
 import ai.koog.agents.core.agent.context.AIAgentContextBase
-import ai.koog.agents.core.agent.entity.AIAgentNodeBase
+import ai.koog.agents.core.agent.entity.graph.AIAgentNodeBase
 
 /**
  * Represents the context for handling node-specific events within the framework.
