+++ agents/agents-test/src/commonMain/kotlin/ai/koog/agents/testing/feature/TestingFeature.kt
@@ -8,6 +8,10 @@ import ai.koog.agents.core.agent.config.AIAgentConfigBase
 import ai.koog.agents.core.agent.context.AIAgentContextBase
 import ai.koog.agents.core.agent.context.AIAgentLLMContext
 import ai.koog.agents.core.agent.entity.*
+import ai.koog.agents.core.agent.entity.graph.AIAgentNodeBase
+import ai.koog.agents.core.agent.entity.graph.AIAgentSubgraph
+import ai.koog.agents.core.agent.entity.graph.FinishNode
+import ai.koog.agents.core.agent.entity.graph.AIAgentGraphStrategy
 import ai.koog.agents.core.annotation.InternalAgentsApi
 import ai.koog.agents.core.environment.AIAgentEnvironment
 import ai.koog.agents.core.environment.ReceivedToolResult
@@ -45,9 +49,9 @@ public sealed class NodeReference<Input, Output> {
     /**
      * Resolves the current node reference within the context of the provided AI agent subgraph.
      *
-     * @param subgraph An instance of [AIAgentSubgraph], representing the structured subgraph within the AI agent workflow.
+     * @param subgraph An instance of [ai.koog.agents.core.agent.entity.graph.AIAgentSubgraph], representing the structured subgraph within the AI agent workflow.
      *                 It contains the logical segment of processing, including the starting and finishing nodes.
-     * @return An instance of [AIAgentNodeBase], representing the resolved node within the subgraph that corresponds
+     * @return An instance of [ai.koog.agents.core.agent.entity.graph.AIAgentNodeBase], representing the resolved node within the subgraph that corresponds
      *         to the current node reference.
      */
     public abstract fun resolve(subgraph: AIAgentSubgraph<*, *>): AIAgentNodeBase<Input, Output>
@@ -178,16 +182,16 @@ public sealed class NodeReference<Input, Output> {
          * @throws IllegalStateException If the subgraph is not of type `AIAgentStrategy`.
          */
         @Suppress("UNCHECKED_CAST")
-        override fun resolve(subgraph: AIAgentSubgraph<*, *>): GraphAIAgentStrategy<Input, Output> {
+        override fun resolve(subgraph: AIAgentSubgraph<*, *>): AIAgentGraphStrategy<Input, Output> {
             if (subgraph.name != name) {
                 throw IllegalArgumentException("Strategy with name '$name' was expected")
             }
 
-            if (subgraph !is GraphAIAgentStrategy) {
+            if (subgraph !is AIAgentGraphStrategy) {
                 throw IllegalStateException("Resolving a strategy is not possible from a subgraph")
             }
 
-            return subgraph as GraphAIAgentStrategy<Input, Output>
+            return subgraph as AIAgentGraphStrategy<Input, Output>
         }
     }
 }
@@ -917,7 +921,7 @@ public class Testing {
      * reachability, outputs, and edges within an AI agent pipeline.
      */
     @TestOnly
-    public companion object Feature : AIAgentFeature<Config, Testing, GraphAIAgentStrategy<*, *>> {
+    public companion object Feature : AIAgentFeature<Config, Testing, AIAgentGraphStrategy<*, *>> {
         /**
          * A storage key uniquely identifying the `Testing` feature within the local agent's storage.
          * The key is generated using the `createStorageKey` function and associates the
@@ -941,7 +945,7 @@ public class Testing {
          */
         override fun install(
             config: Config,
-            pipeline: AIAgentPipeline<out GraphAIAgentStrategy<*, *>>
+            pipeline: AIAgentPipeline<out AIAgentGraphStrategy<*, *>>
         ) {
             val feature = Testing()
             val interceptContext = InterceptContext(this, feature)
@@ -1282,7 +1286,7 @@ public fun toolResult(tool: SimpleTool<*>, result: String): ReceivedToolResult =
  * @see Testing
  * @see Testing.Config
  */
-public fun FeatureContext<out GraphAIAgentStrategy<*, *>>.withTesting(config: Testing.Config.() -> Unit = {}) {
+public fun FeatureContext<out AIAgentGraphStrategy<*, *>>.withTesting(config: Testing.Config.() -> Unit = {}) {
     install(Testing) {
         config()
     }
