+++ examples/src/main/kotlin/ai/koog/agents/example/moderation/JokesWithModeration.kt
@@ -1,7 +1,7 @@
 package ai.koog.agents.example.moderation
 
 import ai.koog.agents.core.agent.AIAgent
-import ai.koog.agents.core.agent.entity.AIAgentNodeBase
+import ai.koog.agents.core.agent.entity.graph.AIAgentNodeBase
 import ai.koog.agents.core.dsl.builder.forwardTo
 import ai.koog.agents.core.dsl.builder.strategy
 import ai.koog.agents.core.dsl.extension.ModeratedMessage
